+++ packages/codegen-ui-react/lib/forms/react-form-renderer.ts
@@ -47,7 +47,7 @@ import {
   SyntaxKind,
   TypeAliasDeclaration,
 } from 'typescript';
-import { lowerCaseFirst } from '../helpers';
+import { buildUseStateExpression, lowerCaseFirst } from '../helpers';
 import { ImportCollection, ImportSource, ImportValue } from '../imports';
 import { PrimitiveTypeParameter, Primitive, primitiveOverrideProp } from '../primitive';
 import { getComponentPropName } from '../react-component-render-helper';
@@ -78,7 +78,6 @@ import {
   getPropName,
 } from './form-renderer-helper';
 import {
-  buildUseStateExpression,
   getCurrentDisplayValueName,
   getArrayChildRefName,
   getCurrentValueName,
@@ -493,12 +492,7 @@ export abstract class ReactFormTemplateRenderer extends StudioTemplateRenderer<
         const destructuredPrimaryKey = getPropName(this.primaryKeys[0]);
         statements.push(
           addUseEffectWrapper(
-            buildUpdateDatastoreQuery(
-              modelName,
-              lowerCaseDataTypeNameRecord,
-              relatedModelStatements,
-              destructuredPrimaryKey,
-            ),
+            buildUpdateDatastoreQuery(modelName, lowerCaseDataTypeName, relatedModelStatements, destructuredPrimaryKey),
             [destructuredPrimaryKey, lowerCaseDataTypeName],
           ),
         );
@@ -607,23 +601,11 @@ export abstract class ReactFormTemplateRenderer extends StudioTemplateRenderer<
     // timestamp type takes precedence over datetime as it includes formatter for datetime
     // we include both the timestamp conversion and local date formatter
     if (dataTypesMap.AWSTimestamp) {
-      // helper needed if update form
-      // or, if create form and the field is an array and therefore controlled
-      if (
-        formActionType === 'update' ||
-        dataTypesMap.AWSTimestamp.some((fieldName) => formMetadata.fieldConfigs[fieldName].isArray)
-      ) {
-        statements.push(convertTimeStampToDateAST, convertToLocalAST);
-      }
+      statements.push(convertTimeStampToDateAST, convertToLocalAST);
     }
     // if we only have date time then we only need the local conversion
     else if (dataTypesMap.AWSDateTime) {
-      if (
-        formActionType === 'update' ||
-        dataTypesMap.AWSDateTime.some((fieldName) => formMetadata.fieldConfigs[fieldName].isArray)
-      ) {
-        statements.push(convertToLocalAST);
-      }
+      statements.push(convertToLocalAST);
     }
 
     return statements;